requirements.txt
Django==1.11.5
djangorestframework==3.7.1
pytz==2017.2
